requirements.txt
python-dotenv
langchain
langchain-openai
openai-whisper
langchain-community
langchain-anthropic
fastapi
uvicorn
pytubefix